telegram_pnl/requirements.txt
ccxt==2.0.71
dataframe-image==0.1.1
matplotlib==3.5.2
pandas==1.4.2
psycopg2-binary==2.9.3
python-telegram-bot==13.12
PyYAML==6.0 